pr #1349. Tests now working.

+++ src/psyclone/tests/test_files/dynamo0p3/15.21.4_int_inc_a_plus_X_builtin.f90
@@ -32,6 +32,7 @@
 ! POSSIBILITY OF SUCH DAMAGE.
 ! -----------------------------------------------------------------------------
 ! Author: I. Kavcic, Met Office
+! Modified: R. W. Ford, STFC Daresbury Lab
 
 program single_invoke
 
@@ -43,7 +44,7 @@ program single_invoke
   implicit none
 
   type(integer_field_type) :: f1
-  real(i_def)              :: a
+  integer(i_def)           :: a
 
   a = 2
